requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.1
flake8-polyfill==1.0.2
idna==2.10
mccabe==0.6.1
pep8-naming==0.11.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
